DVO-948: added images

+++ scripts/bash/download.sh
@@ -15,7 +15,9 @@ Usage:	download [SERVICES] \n\n
 It is going to download bioregister.groovy and apply it to running container. \n
 This script will not download and redeploy bioregister war file. \n\n
 Servics: \n
-\tbioregister   \t    download bioregister.groovy from S3 to running container \n
+\tbioregister -   download bioregister.groovy from S3 to running container \n
+\tsso         -   download sso files from S3 to webapps/browser/WEB-INF/ in running container \n
+\timages      -   download image files from S3 to webapps/browser/images/ in running container \n
 
 EOF
 )
@@ -72,7 +74,7 @@ trap "rm -f $LOCK_FILE" EXIT
 touch $LOCK_FILE
 
 
-# --- Function --------------------------------------------------------
+# --- Functions --------------------------------------------------------
 function download_bioregister_groovy(){
 
     echo "Update Bioregister Groovy:"
@@ -82,21 +84,42 @@ function download_bioregister_groovy(){
 function download_sso_files(){
     echo "Downloading sso files from S3 to browser container ..."
     aws s3 ls s3://$P_INSTALL_BUCKET_NAME/$P_INSTALL_BUCKET_PREFIX/browser/WEB-INF/
-    aws s3 sync  s3://$P_INSTALL_BUCKET_NAME/$P_INSTALL_BUCKET_PREFIX/browser/WEB-INF/ /efs/data/browser/WEB-INF/ --exclude "*.*" --include "sso.*"
+    aws s3 sync  s3://$P_INSTALL_BUCKET_NAME/$P_INSTALL_BUCKET_PREFIX/browser/WEB-INF/ $EFS_CUSTOMED_BROWSER_DIR/WEB-INF/ --exclude "*.*" --include "sso.*"
+    chown -R 1000:1000 $EFS_CUSTOMED_BROWSER_DIR/WEB-INF/
     export BROWSER_CONTAINER_ID=$( docker ps --filter label=app.name=browser --format {{.ID}} )
-    echo "BROWSER_CONTAINER_ID=$BROWSER_CONTAINER_ID"
-    FILES=/efs/data/browser/WEB-INF/sso.*
+    FILES=$EFS_CUSTOMED_BROWSER_DIR/WEB-INF/sso.*
 
     for f in $FILES
     do
-      echo "Processing $f file..."
-      docker cp $f $BROWSER_CONTAINER_ID:/usr/local/tomcat/webapps/browser/WEB-INF/
-      docker exec -t $BROWSER_CONTAINER_ID  chown -R 1000:1000 /usr/local/tomcat/webapps/browser/WEB-INF/
+      if [   -f "$f" ]; then
+          echo "Processing $f file..."
+          docker cp $f $BROWSER_CONTAINER_ID:/usr/local/tomcat/webapps/browser/WEB-INF/
+          docker exec -t $BROWSER_CONTAINER_ID chown -R 1000:1000 /usr/local/tomcat/webapps/browser/WEB-INF/
+      fi
     done
 
-    echo "All files in browser/WEB-INF/"
-    docker exec -t $BROWSER_CONTAINER_ID  ls -lsa /usr/local/tomcat/webapps/browser/WEB-INF/
-    echo -e "\nAll sso files have been copied from S3 to browser/WEB-INF/"
+    echo -e "\ndownload_sso_files done"
+}
+
+
+function download_image_files(){
+    echo "Downloading image files from S3 to browser container ..."
+    aws s3 ls s3://$P_INSTALL_BUCKET_NAME/$P_INSTALL_BUCKET_PREFIX/browser/images/
+    aws s3 sync  s3://$P_INSTALL_BUCKET_NAME/$P_INSTALL_BUCKET_PREFIX/browser/images/ $EFS_CUSTOMED_BROWSER_DIR/images/  --include "*.*"
+    chown -R 1000:1000 $EFS_CUSTOMED_BROWSER_DIR/images/
+    export BROWSER_CONTAINER_ID=$( docker ps --filter label=app.name=browser --format {{.ID}} )
+    FILES=$EFS_CUSTOMED_BROWSER_DIR/images/images/*
+
+    for f in $FILES
+    do
+      if [   -f "$f" ]; then
+          echo "Processing $f file..."
+          docker cp $f $BROWSER_CONTAINER_ID:/usr/local/tomcat/webapps/browser/images/
+          docker exec -t $BROWSER_CONTAINER_ID chown -R 1000:1000 /usr/local/tomcat/webapps/browser/images/
+      fi
+    done
+
+    echo -e "\ndownload_image_files done"
 }
 
 function check_env_configuration(){
@@ -123,6 +146,10 @@ elif [ -z "$PRIVATE_DNS_NAME" ]; then
 
 elif [ -z "$P_DATABASE_NAME" ]; then
     echo "[ERROR] P_DATABASE_NAME env variable cannot be empty."
+    exit 0 ;
+
+elif [ -z "$EFS_CUSTOMED_BROWSER_DIR" ]; then
+    echo "[ERROR] EFS_CUSTOMED_BROWSER_DIR env variable cannot be empty."
     exit 0 ;
 
 
@@ -152,6 +179,10 @@ if [ "bioregister" = $param1 ]; then
 
 elif [ "sso" = $param1 ]; then
     download_sso_files
+
+
+elif [ "images" = $param1 ]; then
+    download_image_files
 
 else
     echo "invalid service '$param1'"
